web_fastapi/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyMySQL==1.1.0
redis==5.0.1
pymongo==4.6.0
python-multipart==0.0.6
